+++ Controllers/ProjectLinksController.cs
@@ -1,8 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using GymLinkPro.Data;
@@ -43,7 +42,6 @@ namespace GymLinkPro.Controllers
         }
 
         // PUT: api/ProjectLinks/5
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
         public async Task<IActionResult> PutProjectLink(int id, ProjectLink projectLink)
         {
@@ -74,14 +72,15 @@ namespace GymLinkPro.Controllers
         }
 
         // POST: api/ProjectLinks
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
         public async Task<ActionResult<ProjectLink>> PostProjectLink(ProjectLink projectLink)
         {
+            projectLink.AddedByUserId = GetCurrentUserId();
+
             _context.ProjectLinks.Add(projectLink);
             await _context.SaveChangesAsync();
 
-            return CreatedAtAction("GetProjectLink", new { id = projectLink.ProjectLinkId }, projectLink);
+            return CreatedAtAction(nameof(GetProjectLink), new { id = projectLink.ProjectLinkId }, projectLink);
         }
 
         // DELETE: api/ProjectLinks/5
@@ -94,6 +93,21 @@ namespace GymLinkPro.Controllers
                 return NotFound();
             }
 
+            int currentUserId = GetCurrentUserId();
+
+            bool isOwner = projectLink.AddedByUserId == currentUserId;
+
+            var membership = await _context.ProjectMembers
+                .FirstOrDefaultAsync(pm => pm.ProjectId == projectLink.ProjectId && pm.MemberId == currentUserId);
+
+            bool isAdminOrCoAdmin = membership != null &&
+                (membership.Role == "Admin" || membership.Role == "Co-Admin");
+
+            if (!isOwner && !isAdminOrCoAdmin)
+            {
+                return Forbid();
+            }
+
             _context.ProjectLinks.Remove(projectLink);
             await _context.SaveChangesAsync();
 
@@ -103,6 +117,16 @@ namespace GymLinkPro.Controllers
         private bool ProjectLinkExists(int id)
         {
             return _context.ProjectLinks.Any(e => e.ProjectLinkId == id);
+        }
+
+        private int GetCurrentUserId()
+        {
+            var userIdClaim = User.FindFirst(ClaimTypes.NameIdentifier);
+            if (userIdClaim != null && int.TryParse(userIdClaim.Value, out int userId))
+            {
+                return userId;
+            }
+            throw new UnauthorizedAccessException("User ID not found in claims.");
         }
     }
 }
